requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==3.0.4
discord.py==1.7.3
idna==2.10
multidict==6.0.2
python-dotenv==0.20.0
typing_extensions==4.3.0
yarl==1.7.2
openai==0.26.5
pytz==2022.5
googletrans==3.1.0a0
